Update correlation.py
i changed so much. pretty much fixed it so it actually prints the essid of the correlated mac next to the mac as long as it is the same in both csvs. also chose to use csv.reader instead of csv.dictreader because of the problems i kept running into with trying to use dictreader.